Create 'LayoutController' component:
* Create the componenent in 'client' package (as a single source of truth for the app layout
* Use the controller on each page (passing in only the page-specific content) instead of the 'Layout' from the UI
