--- packages/client/src/data/navigation.ts
@@ -1,27 +1,0 @@
-import { LinkItem } from "@eisbuk/ui";
-import { PrivateRoutes } from "@eisbuk/shared/ui";
-import i18n, { NavigationLabel } from "@eisbuk/translations";
-import { ClipboardList, Users, DocumentDuplicate, Cog } from "@eisbuk/svg";
-
-export const adminLinks: LinkItem[] = [
-  {
-    Icon: ClipboardList,
-    label: i18n.t(NavigationLabel.Attendance),
-    slug: PrivateRoutes.Root,
-  },
-  {
-    Icon: DocumentDuplicate,
-    label: i18n.t(NavigationLabel.Slots),
-    slug: PrivateRoutes.Slots,
-  },
-  {
-    Icon: Users,
-    label: i18n.t(NavigationLabel.Athletes),
-    slug: PrivateRoutes.Athletes,
-  },
-  {
-    Icon: Cog,
-    label: i18n.t(NavigationLabel.OrganizationSettings),
-    slug: PrivateRoutes.AdminPreferences,
-  },
-];
