+++ packages/client/src/pages/attendance/index.tsx
@@ -4,13 +4,7 @@ import { Link } from "react-router-dom";
 import { DateTime } from "luxon";
 import { useDispatch, useSelector } from "react-redux";
 
-import {
-  CalendarNav,
-  Layout,
-  TabItem,
-  CalendarNavProps,
-  Button,
-} from "@eisbuk/ui";
+import { CalendarNav, TabItem, CalendarNavProps, Button } from "@eisbuk/ui";
 import { Calendar, Printer } from "@eisbuk/svg";
 import { OrgSubCollection } from "@eisbuk/shared";
 import i18n, { AttendanceNavigationLabel } from "@eisbuk/translations";
@@ -21,14 +15,11 @@ import { getOrganization } from "@/lib/getters";
 import ByDayView from "./views/ByDay";
 import ByMonthView from "./views/ByMonth";
 
-import { NotificationsContainer } from "@/features/notifications/components";
-import BirthdayMenu from "@/controllers/BirthdayMenu";
+import Layout from "@/controllers/Layout";
 
 import { getCalendarDay } from "@/store/selectors/app";
 
 import { changeCalendarDate } from "@/store/actions/appActions";
-
-import { adminLinks } from "@/data/navigation";
 
 enum Views {
   ByDay = "ByDayView",
@@ -81,8 +72,6 @@ const AttendancePage: React.FC = () => {
       </Link>
     ) : undefined;
 
-  const additionalAdminContent = <BirthdayMenu />;
-
   const additionalButtons = (
     <>
       <TabItem
@@ -103,13 +92,7 @@ const AttendancePage: React.FC = () => {
   );
 
   return (
-    <Layout
-      isAdmin
-      adminLinks={adminLinks}
-      Notifications={NotificationsContainer}
-      additionalAdminContent={additionalAdminContent}
-      additionalButtons={additionalButtons}
-    >
+    <Layout additionalButtons={additionalButtons}>
       <CalendarNav
         {...calendarNavProps}
         jump={calendarJump}
