+++ packages/client/src/pages/customer_area/index.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import { useSelector } from "react-redux";
 import i18n, { CustomerNavigationLabel } from "@eisbuk/translations";
 
-import { CalendarNav, Layout, TabItem } from "@eisbuk/ui";
+import { CalendarNav, TabItem } from "@eisbuk/ui";
 import { Calendar, AccountCircle, ClipboardList } from "@eisbuk/svg";
 import {
   BookingSubCollection,
@@ -18,15 +18,11 @@ import CalendarView from "./views/Calendar";
 import ProfileView from "./views/Profile";
 import { useSecretKey, useDate } from "./hooks";
 
-import { NotificationsContainer } from "@/features/notifications/components";
 import AddToCalendar from "@/components/atoms/AddToCalendar";
 
-import BirthdayMenu from "@/controllers/BirthdayMenu";
+import Layout from "@/controllers/Layout";
 
 import { getBookingsCustomer } from "@/store/selectors/bookings";
-import { getIsAdmin } from "@/store/selectors/auth";
-
-import { adminLinks } from "@/data/navigation";
 
 enum Views {
   Book = "BookView",
@@ -46,10 +42,6 @@ const viewsLookup = {
  */
 const CustomerArea: React.FC = () => {
   const secretKey = useSecretKey();
-
-  const isAdmin = useSelector(getIsAdmin);
-
-  const additionalAdminContent = <BirthdayMenu />;
 
   // Subscribe to necessary collections
   useFirestoreSubscribe(getOrganization(), [
@@ -96,14 +88,7 @@ const CustomerArea: React.FC = () => {
   );
 
   return (
-    <Layout
-      isAdmin={isAdmin}
-      adminLinks={adminLinks}
-      Notifications={NotificationsContainer}
-      additionalButtons={additionalButtons}
-      additionalAdminContent={additionalAdminContent}
-      user={userData}
-    >
+    <Layout additionalButtons={additionalButtons} user={userData}>
       {view !== "ProfileView" && (
         <CalendarNav
           {...calendarNavProps}
