+++ packages/client/src/pages/customers/index.tsx
@@ -4,7 +4,7 @@ import { Link, useHistory } from "react-router-dom";
 
 import { OrgSubCollection, Customer } from "@eisbuk/shared";
 import { PrivateRoutes } from "@eisbuk/shared/ui";
-import { Layout, CustomerGrid, SearchBar } from "@eisbuk/ui";
+import { CustomerGrid, SearchBar } from "@eisbuk/ui";
 import {
   useTranslation,
   NavigationLabel,
@@ -15,8 +15,7 @@ import { useFirestoreSubscribe } from "@eisbuk/react-redux-firebase-firestore";
 
 import { __addAthleteId__ } from "@eisbuk/testing/testIds";
 
-import BirthdayMenu from "@/controllers/BirthdayMenu";
-import { NotificationsContainer } from "@/features/notifications/components";
+import Layout from "@/controllers/Layout";
 
 import { getCustomersList } from "@/store/selectors/customers";
 
@@ -24,8 +23,6 @@ import useTitle from "@/hooks/useTitle";
 
 import { isEmpty } from "@/utils/helpers";
 import { getOrganization } from "@/lib/getters";
-
-import { adminLinks } from "@/data/navigation";
 
 const AthletesPage: React.FC = () => {
   const { t } = useTranslation();
@@ -36,8 +33,6 @@ const AthletesPage: React.FC = () => {
   ]);
 
   useTitle(t(NavigationLabel.Athletes));
-
-  const additionalAdminContent = <BirthdayMenu />;
 
   const customers = useSelector(getCustomersList(true));
 
@@ -50,12 +45,7 @@ const AthletesPage: React.FC = () => {
 
   /** @TODO update below when we create `isEmpty` and `isLoaded` helpers */
   return (
-    <Layout
-      isAdmin
-      adminLinks={adminLinks}
-      Notifications={NotificationsContainer}
-      additionalAdminContent={additionalAdminContent}
-    >
+    <Layout>
       <div className="content-container !pt-16">
         {!isEmpty(customers) && (
           <>
