+++ packages/client/src/pages/slots/index.tsx
@@ -11,7 +11,6 @@ import {
   Button,
   ButtonColor,
   CalendarNav,
-  Layout,
   SlotsDayContainer,
 } from "@eisbuk/ui";
 
@@ -24,14 +23,13 @@ import { LocalStore, SlotsWeek } from "@/types/store";
 
 import { getOrganization } from "@/lib/getters";
 
+import Layout from "@/controllers/Layout";
 import SlotOperationButtons, {
   CopyButton,
   PasteButton,
   NewSlotButton,
 } from "@/components/atoms/SlotOperationButtons";
 import SlotCard from "@/controllers/SlotCard";
-import BirthdayMenu from "@/controllers/BirthdayMenu";
-import { NotificationsContainer } from "@/features/notifications/components";
 
 import { getAdminSlots } from "@/store/selectors/slots";
 import { getCalendarDay } from "@/store/selectors/app";
@@ -47,8 +45,6 @@ import {
 import { changeCalendarDate } from "@/store/actions/appActions";
 
 import { comparePeriods } from "@/utils/sort";
-
-import { adminLinks } from "@/data/navigation";
 
 const SlotsPage: React.FC = () => {
   const dispatch = useDispatch();
@@ -70,8 +66,6 @@ const SlotsPage: React.FC = () => {
   const attendance = useSelector(
     (state: LocalStore) => state.firestore.data.attendance || {}
   );
-
-  const additionalAdminContent = <BirthdayMenu />;
 
   const { t } = useTranslation();
 
@@ -130,12 +124,7 @@ const SlotsPage: React.FC = () => {
     };
 
   return (
-    <Layout
-      isAdmin
-      adminLinks={adminLinks}
-      Notifications={NotificationsContainer}
-      additionalAdminContent={additionalAdminContent}
-    >
+    <Layout>
       <CalendarNav
         date={currentDate}
         onChange={(date) => dispatch(changeCalendarDate(date))}
